Refactor

+++ zal/info_rejestrator/main.c
@@ -87,7 +87,7 @@ int parseParameters(int argc, char* argv[], int* signalNo, int* pid) {
 int StringToInt(char *string, int *wasErr) {
     char *endptr = NULL;
     errno = 0;
-    int val = strtol(string, &endptr, 10);
+    long val = strtol(string, &endptr, 10);
     if ((errno == ERANGE && (val == LONG_MAX || val == LONG_MIN))
         || (errno != 0 && val == 0)
         || endptr == string
@@ -95,5 +95,5 @@ int StringToInt(char *string, int *wasErr) {
     ) {
         *wasErr = -1;
     }
-    return val;
+    return (int)val;
 }
